Penalty refactoring.
