+++ src/main/scala/scorex/core/network/NodeViewSynchronizer.scala
@@ -9,7 +9,6 @@ import scorex.core.NodeViewHolder.ReceivableMessages.{GetNodeViewChanges, Modifi
 import scorex.core.consensus.History._
 import scorex.core.consensus.{History, HistoryReader, SyncInfo}
 import scorex.core.network.ModifiersStatus.Requested
-import scorex.core.network.NetworkController.PenaltyType
 import scorex.core.network.NetworkController.ReceivableMessages.{BlacklistPeer, RegisterMessageSpecs, SendToNetwork}
 import scorex.core.network.NetworkControllerSharedMessages.ReceivableMessages.DataFromPeer
 import scorex.core.network.NodeViewSynchronizer.ReceivableMessages._
