+++ src/main/scala/scorex/core/network/PenaltyType.scala
@@ -1,0 +1,9 @@
+package scorex.core.network
+
+sealed trait PenaltyType
+
+object PenaltyType {
+  case object MisbehaviorPenalty extends PenaltyType
+  case object SpamPenalty extends PenaltyType
+  case object PermanentPenalty extends PenaltyType
+}
